+++ packages/cli/src/publish.ts
@@ -5,7 +5,7 @@ import * as api from '@gitbook/api';
 
 import { buildScriptFromManifest } from './build';
 import { resolveFile } from './manifest';
-import { assertCanPublishIntegrations, getAPIClient } from './remote';
+import { getAPIClient } from './remote';
 
 /**
  * Publish the integration to GitBook.
@@ -15,12 +15,10 @@ export async function publishIntegration(
     specFilePath: string,
     updates: Partial<api.RequestPublishIntegration> = {},
 ): Promise<void> {
-    assertCanPublishIntegrations();
-
     // Build the script
     const { script, manifest } = await buildScriptFromManifest(specFilePath);
 
-    const api = await getAPIClient(true, { personalTokenOnly: true });
+    const api = await getAPIClient(true);
 
     if (typeof manifest.target === 'string') {
         console.log(
@@ -65,9 +63,7 @@ export async function publishIntegration(
  * Delete an integration
  */
 export async function unpublishIntegration(name: string): Promise<void> {
-    assertCanPublishIntegrations();
-
-    const api = await getAPIClient(true, { personalTokenOnly: true });
+    const api = await getAPIClient(true);
     await api.integrations.unpublishIntegration(name);
 
     console.log(`👌 Integration "${name}" has been deleted`);